requirements.txt
Flask
pyTelegramBotAPI
gunicorn
google-cloud-functions
google-cloud-pubsub	
google-cloud-storage
python-dotenv
openpyxl
vertexai
pillow
typing